entrega3_impl/requirements.txt
flask
psycopg2-binary
lxml
